services/vanna/requirements.txt
fastapi
uvicorn
psycopg[binary]
sqlalchemy
groq
pandas
python-dotenv
openai